% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc_cna}
\alias{acc_cna}
\title{Copy number alteration data from ACC TCGA}
\format{
A data frame with 23878 rows of CNA data from
}
\usage{
acc_cna
}
\description{
A dataset of CNA, sourced from cBioPortal
}
\keyword{datasets}
