% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acc_mutation}
\alias{acc_mutation}
\title{Mutation data from ACC TCGA}
\format{
A data frame with 13486 rows of mutation data
}
\usage{
acc_mutation
}
\description{
A dataset of mutations, sourced from cBioPortal
}
\keyword{datasets}
