% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cna.R
\name{annotate_cna}
\alias{annotate_cna}
\title{Annotate CNA}
\usage{
annotate_cna(
  cna,
  annotate_tumor_type = NULL,
  return_simple = TRUE,
  return_query_params = FALSE,
  reference_genome = "GRCh37",
  token = get_oncokb_token()
)
}
\arguments{
\item{cna}{a cna file in long format (similar to maf file)}

\item{return_simple}{Default is \code{TRUE} where only a set of the most common columns are returned from oncoKB annotator
see \code{oncokbR::output_dictionary} for more information on what is returned. If \code{FALSE} all raw columns are returned from API.}

\item{return_query_params}{If \code{TRUE}, the specific parameters used to query the API are returned in new columns.
This can be useful for troubleshooting the annotator. Default is \code{FALSE}.}
}
\value{
an annotated cna file
}
\description{
Annotate CNA
}
\examples{
ex_cna <- blca_cna[1:10,]
ex_cna$tumor_type = "BLCA"

x <- annotate_cna(ex_cna)

}
