% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutations.R
\name{annotate_mutations}
\alias{annotate_mutations}
\title{Annotate Mutations}
\usage{
annotate_mutations(mutations, annotate_by = c("protein_change", "hgvsg"))
}
\arguments{
\item{mutations}{a mutations file in MAF, or similar format}

\item{annotate_by}{Can indicate whether to annotate by "protein_change" or "hgvsg (see oncoKB API docs for more info).
"Default is \code{protein_change}}
}
\value{
an annotated mutations file
}
\description{
Annotate maf and maf-like files using information about protein change or HGVSg.
You can annotated with tumor types indicated (oncotree codes), or without.
}
\examples{
mutations <- oncokbR::blca_mutation[1:10, ]

# Annotate without tumor type -----
x <- annotate_mutations(mutations)

# Annotate with tumor type -----

# Annotate by HGVSg
z <- annotate_mutations(oncokbR::mutations_hgvsg, annotate_by =  "hgvsg")

}
