% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sv.R
\name{annotate_sv}
\alias{annotate_sv}
\title{Annotate Structural Variants}
\usage{
annotate_sv(sv)
}
\arguments{
\item{sv}{a cna file in long format (similar to maf file)}
}
\value{
an annotated sv file
}
\description{
Annotate Structural Variants
}
\examples{
sv <- blca_sv[1:10,]

x <- annotate_sv(sv = sv)

}
