% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blca_cna}
\alias{blca_cna}
\title{Copy number alteration data from Nonmuscle invasive bladder cancer patients (MSK Eur Urol 2017)}
\format{
A data frame with 168 rows of CNA data from
Nonmuscle invasive bladder cancer patients (MSK Eur Urol 2017)
\describe{
\item{hugoGeneSymbol}{Hugo Gene Symbol}
}
}
\usage{
blca_cna
}
\description{
A dataset of CNA, sourced from cBioPortal
}
\keyword{datasets}
