% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blca_mutation}
\alias{blca_mutation}
\title{Mutation data from Nonmuscle invasive bladder cancer patients (MSK Eur Urol 2017)}
\format{
A data frame with 1595 rows of mutation data from
Nonmuscle invasive bladder cancer patients (MSK Eur Urol 2017)
\describe{
\item{hugoGeneSymbol}{Hugo Gene Symbol}
}
}
\usage{
blca_mutation
}
\description{
A dataset of mutations, sourced from cBioPortal
}
\keyword{datasets}
