% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blca_sv}
\alias{blca_sv}
\title{Mutation data from Nonmuscle invasive bladder cancer patients (MSK Eur Urol 2017)}
\format{
A data frame with 13 rows of structural variant data from
}
\usage{
blca_sv
}
\description{
A dataset of mutations, sourced from cBioPortal
}
\keyword{datasets}
