% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutations.R
\name{.annotate_mutations_by_hgvsg}
\alias{.annotate_mutations_by_hgvsg}
\title{Annotate Mutations By HGVSg}
\usage{
.annotate_mutations_by_hgvsg(mutations, annotate_tumor_type)
}
\arguments{
\item{mutations}{}
}
\value{
a dataframe of annotated mutations
}
\description{
Annotate Mutations By HGVSg
}
\examples{


mutations <- rename_columns(oncokbR::mutations_hgvsg)
mutations$index = 1:nrow(mutations)
.annotate_mutations_by_hgvsg(mutations, annotate_tumor_type = FALSE)

}
\keyword{internal}
