% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutations.R
\name{.annotate_mutations_by_protein_change}
\alias{.annotate_mutations_by_protein_change}
\title{Annotate Mutations By Protein Change}
\usage{
.annotate_mutations_by_protein_change(mutations, annotate_tumor_type)
}
\arguments{
\item{mutations}{}
}
\value{
a dataframe of annotated mutations
}
\description{
Annotate Mutations By Protein Change
}
\examples{
mutations <- rename_columns(oncokbR::blca_mutation[1:10, ])
.annotate_mutations_by_protein_change(mutations)

}
\keyword{internal}
