% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.clean_query_results}
\alias{.clean_query_results}
\title{Clean API Query Results}
\usage{
.clean_query_results(
  query_result,
  original_data,
  return_simple,
  return_query_params
)
}
\arguments{
\item{query_result}{a data frame returned from API}

\item{original_data}{the original user input data frame}

\item{return_simple}{Default is \code{TRUE} where only a set of the most common columns are returned from oncoKB annotator
see \code{oncokbR::output_dictionary} for more information on what is returned. If \code{FALSE} all raw columns are returned from API.}

\item{return_query_params}{If \code{TRUE}, the specific parameters used to query the API are returned in new columns.
This can be useful for troubleshooting the annotator. Default is \code{FALSE}.}
}
\value{
a cleaned data frame
}
\description{
Clean API Query Results
}
