% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_select_columns}
\alias{.get_select_columns}
\title{Helper to select columns for output of annotation functions}
\usage{
.get_select_columns(results, return_simple, return_query_params)
}
\arguments{
\item{results}{input annotation result data frame}

\item{return_simple}{Default is \code{TRUE} where only a set of the most common columns are returned from oncoKB annotator
see \code{oncokbR::output_dictionary} for more information on what is returned. If \code{FALSE} all raw columns are returned from API.}

\item{return_query_params}{If \code{TRUE}, the specific parameters used to query the API are returned in new columns.
This can be useful for troubleshooting the annotator. Default is \code{FALSE}.}
}
\value{
a string of select column names
}
\description{
Helper to select columns for output of annotation functions
}
