% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tumor_type_warning}
\alias{.tumor_type_warning}
\title{Warn if no tumor type annotation possible}
\usage{
.tumor_type_warning(annotate_tumor_type, data)
}
\arguments{
\item{annotate_tumor_type}{TRUE if tumor type column is available in data}

\item{data}{mutation, cna or sv data}
}
\value{
a warning and data set with edited columns
}
\description{
Warn if no tumor type annotation possible
}
