% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_oncokb_token}
\alias{get_oncokb_token}
\title{Get oncoKB Access Token}
\usage{
get_oncokb_token(token = NULL)
}
\value{
Returns a string with token if successfully authenticated, or
a warning that token could not be found.
}
\description{
Convenience function that retrieves oncoKB token System Environment variable "ONCOKB_TOKEN"
}
\examples{
\dontrun{
get_oncokb_token()
}

}
\author{
Karissa Whiting
}
