% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutations_hgvsg}
\alias{mutations_hgvsg}
\title{Generated mutation data}
\format{
A data frame with 13 rows of structural variant data from
}
\usage{
mutations_hgvsg
}
\description{
A dataset of mutations, with hgvsg data
}
\keyword{datasets}
