% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_df}
\alias{names_df}
\title{Column Names Dictionary}
\format{
A data frame of accepted column names
}
\usage{
names_df
}
\description{
Dataset of recoding values for consequence mutation data
}
\keyword{datasets}
