% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokbR-package.R
\docType{package}
\name{oncokbR-package}
\alias{oncokbR}
\alias{oncokbR-package}
\title{oncokbR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
