% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_dictionary}
\alias{output_dictionary}
\title{Output Dictionary}
\format{
A data frame with one row per returned data pint
}
\usage{
output_dictionary
}
\description{
describes output columns returned from oncoKB API
}
\keyword{datasets}
