% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns from API results to work with gnomeR functions}
\usage{
rename_columns(df_to_check)
}
\arguments{
\item{df_to_check}{a data frame for which to check colums}
}
\value{
a renamed data frame
}
\description{
Rename columns from API results to work with gnomeR functions
}
\examples{

rename_columns(df_to_check = oncokbR::blca_mutation)
rename_columns(df_to_check = oncokbR::blca_sv)

}
