% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cna.R
\name{annotate_cna}
\alias{annotate_cna}
\title{Annotate CNA}
\usage{
annotate_cna(cna)
}
\arguments{
\item{cna}{a cna file in long format (similar to maf file)}
}
\value{
an annotated cna file
}
\description{
Annotate CNA
}
\examples{
ex_cna <- blca_cna[1:10,]
ex_cna$tumor_type = "BLCA"

x <- annotate_cna(ex_cna)

}
