% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutations.R
\name{annotate_mutations}
\alias{annotate_mutations}
\title{Annotate Mutations}
\usage{
annotate_mutations(mutations)
}
\arguments{
\item{mutations}{a mutations file in MAF, or similar format}
}
\value{
an annotated mutations file
}
\description{
Annotate Mutations
}
\examples{
ex_mut <- blca_mutation[1:50, ]
ex_mut$tumor_type = "BLCA"
annotate_mutations(ex_mut)

}
