% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutations.R
\name{annotate_mutations}
\alias{annotate_mutations}
\title{Annotate Mutations}
\usage{
annotate_mutations(
  mutations,
  annotate_by = c("protein_change", "hgvsg"),
  annotate_tumor_type = NULL,
  return_simple = TRUE,
  return_query_params = FALSE,
  reference_genome = "GRCh37",
  token = get_oncokb_token()
)
}
\arguments{
\item{mutations}{a mutations file in MAF, or similar format}

\item{annotate_by}{Can indicate whether to annotate by "protein_change" or "hgvsg (see oncoKB API docs for more info).
"Default is \code{protein_change}}

\item{return_simple}{Default is \code{TRUE} where only a set of the most common columns are returned from oncoKB annotator
see \code{oncokbR::output_dictionary} for more information on what is returned. If \code{FALSE} all raw columns are returned from API.}

\item{return_query_params}{If \code{TRUE}, the specific parameters used to query the API are returned in new columns.
This can be useful for troubleshooting the annotator. Default is \code{FALSE}.}
}
\value{
an annotated mutations file
}
\description{
Annotate maf and maf-like files using information about protein change or HGVSg.
You can annotated with tumor types indicated (oncotree codes), or without.
}
\examples{
mutations <- oncokbR::blca_mutation[1:10, ]

# Annotate without tumor type -----
x <- annotate_mutations(mutations)

# Annotate with tumor type -----

# Annotate by HGVSg
z <- annotate_mutations(mutations = oncokbR::mutations_hgvsg, annotate_by =  "hgvsg")
z <- annotate_mutations(mutations = oncokbR::blca_mutation, annotate_by =  "protein_change")

}
