% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokbR-package.R
\docType{package}
\name{oncokbR-package}
\alias{oncokbR}
\alias{oncokbR-package}
\title{oncokbR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.karissawhiting.com/oncokbR/}
}

}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

}
\keyword{internal}
